Fixed data types to avoid potential overflow issues when dealing with strings

+++ string/ft_substr.c
@@ -6,13 +6,13 @@
 /*   By: chlimous <chlimous@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/10/25 22:17:49 by chlimous          #+#    #+#             */
-/*   Updated: 2023/11/21 23:22:43 by chlimous         ###   ########.fr       */
+/*   Updated: 2024/01/07 21:38:36 by chlimous         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-static size_t	substr_size(char const *s, unsigned int start, size_t len)
+static size_t	substr_size(char const *s, size_t start, size_t len)
 {
 	size_t		size;
 
@@ -24,7 +24,7 @@ static size_t	substr_size(char const *s, unsigned int start, size_t len)
 	return (size);
 }
 
-char	*ft_substr(char const *s, unsigned int start, size_t len)
+char	*ft_substr(char const *s, size_t start, size_t len)
 {
 	char		*str;
 	size_t		size;
